+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.39](https://github.com/sting-ioc/sting/tree/v0.39) (2026-06-23) · [Full Changelog](https://github.com/sting-ioc/sting/compare/v0.38...v0.39)
 
 Changes in this release: